+++ gui/ui_randogui.py
@@ -517,6 +517,11 @@ class Ui_MainWindow(object):
 
         self.vlay_dungeons.addLayout(self.vlay_sword_reward)
 
+        self.option_randomize_boss_key_puzzles = QCheckBox(self.verticalLayoutWidget_7)
+        self.option_randomize_boss_key_puzzles.setObjectName(u"option_randomize_boss_key_puzzles")
+
+        self.vlay_dungeons.addWidget(self.option_randomize_boss_key_puzzles)
+
         self.vspace_dungeons = QSpacerItem(20, 40, QSizePolicy.Minimum, QSizePolicy.Expanding)
 
         self.vlay_dungeons.addItem(self.vspace_dungeons)
@@ -1260,7 +1265,7 @@ class Ui_MainWindow(object):
 
         self.retranslateUi(MainWindow)
 
-        self.tabWidget.setCurrentIndex(2)
+        self.tabWidget.setCurrentIndex(1)
         self.option_triforce_shuffle.setCurrentIndex(-1)
         self.option_randomize_entrances.setCurrentIndex(-1)
         self.option_chest_dowsing.setCurrentIndex(-1)
@@ -1323,6 +1328,7 @@ class Ui_MainWindow(object):
         self.label_for_option_boss_key_mode.setText(QCoreApplication.translate("MainWindow", u"Boss Keys", None))
         self.option_empty_unrequired_dungeons.setText(QCoreApplication.translate("MainWindow", u"Empty Unrequired Dungeons", None))
         self.label_for_sword_dungeon_reward.setText(QCoreApplication.translate("MainWindow", u"Sword Dungeon Reward", None))
+        self.option_randomize_boss_key_puzzles.setText(QCoreApplication.translate("MainWindow", u"Randomize Boss Key Puzzles", None))
         self.box_open.setTitle(QCoreApplication.translate("MainWindow", u"Open Settings", None))
         self.label_for_option_open_thunderhead.setText(QCoreApplication.translate("MainWindow", u"Open Thunderhead", None))
         self.option_open_et.setText(QCoreApplication.translate("MainWindow", u"Open Earth Temple", None))