+++ logic/placement_file.py
@@ -24,6 +24,7 @@ class PlacementFile:
         self.trial_connections = {}
         self.trial_object_seed = -1
         self.music_rando_seed = -1
+        self.bk_angle_seed = -1
 
     def read_from_file(self, f):
         self._read_from_json(json.load(f))
@@ -45,6 +46,7 @@ class PlacementFile:
             "trial-connections": self.trial_connections,
             "trial-object-seed": self.trial_object_seed,
             "music-rando-seed": self.music_rando_seed,
+            "bk-angle-seed": self.bk_angle_seed,
         }
         return json.dumps(retval, indent=2)
 
@@ -62,6 +64,7 @@ class PlacementFile:
         self.trial_connections = jsn["trial-connections"]
         self.trial_object_seed = jsn["trial-object-seed"]
         self.music_rando_seed = jsn["music-rando-seed"]
+        self.bk_angle_seed = jsn["bk-angle-seed"]
 
     def check_valid(self, areas):
         """checks, if the current state is valid, throws an exception otherwise